requirements.txt
vgamepad
mss
pillow 
flask 